add pCache to store  to fix startup
